Finish Rebuild
完成重构，现在看起来很精简了，不过灵活性上还欠缺考虑，但是职责分配比较清晰了，下面尝试做一下UI

+++ src/dao/record_dao/RecordDAOImpl.java
@@ -19,68 +19,36 @@ public class RecordDAOImpl extends BaseDao implements RecordDAO {
 
 
     @Override
-    public void addRecord(Record record) {
+    public void addRecord(Record record) throws IllegalArgumentException{
         try {
-            String addSql = "INSERT INTO Record VALUES (null, ?, ?, ?, ?)";
-            Object[] param = new Object[4];
-            param[0] = record.getUserId();
-            param[1] = record.getBookId();
-            param[2] = record.getBorrowDate().toString();
-            param[3] = record.getReturnDate().toString();
-            executeSQL(addSql, param);
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        }
-    }
-
-    @Override
-    public void deleteRecord(int userId, int bookId) {
-        try {
-            String delSql = "DELETE FROM Record WHERE user_id = ? AND book_id = ?";
-            Object[] param = new Object[2];
-            param[0] = userId;
-            param[1] = bookId;
-            executeSQL(delSql, param);
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        }
-    }
-
-    @Override
-    public ArrayList<Record> checkOneRecords(int userId) throws IllegalArgumentException {
-        try {
-            Connection connect = getConnection();
-            String checkSql = "SELECT * FROM Record WHERE user_id = ?";
-            PreparedStatement check = connect.prepareStatement(checkSql);
-            check.setObject(1, userId);
-            ResultSet rst = check.executeQuery();
-            if (rst.next()) {
-                ArrayList<Record> records = new ArrayList<>();
-                int bookId = rst.getInt("book_id");
-                Date borrowDay = new Date(rst.getDate("borrow_day").toString());
-                Date returnDay = new Date(rst.getDate("return_day").toString());
-                records.add(new Record(userId, bookId, borrowDay, returnDay));
-                while (rst.next()) {
-                    bookId = rst.getInt("book_id");
-                    borrowDay = new Date(rst.getDate("borrow_day").toString());
-                    returnDay = new Date(rst.getDate("return_day").toString());
-                    records.add(new Record(userId, bookId, borrowDay, returnDay));
-                }
-                closeAll(connect, check, rst);
-                return records;
+            if (!haveRecord(record.getUserId(), record.getBookId())) {
+                String addSql = "INSERT INTO Record VALUES (null, ?, ?, ?, ?)";
+                executeSQL(addSql, record.getUserId(), record.getBookId(), record.getBorrowDate().toString(), record.getReturnDate().toString());
             }
             else {
-                closeAll(connect, check, rst);
-                throw new IllegalArgumentException("No Borrow Record!");
+                throw new IllegalArgumentException("You have Borrowed This Book!");
             }
-        } catch (SQLException e) {
+        } catch (ClassNotFoundException e) {
             e.printStackTrace();
         }
-        return null;
     }
 
     @Override
-    public boolean hadBorrow(int userId, int bookId) {
+    public void deleteRecord(int userId, int bookId) throws IllegalArgumentException {
+        try {
+            if (haveRecord(userId, bookId)) {
+                String delSql = "DELETE FROM Record WHERE user_id = ? AND book_id = ?";
+                executeSQL(delSql, userId, bookId);
+            }
+            else {
+                throw new IllegalArgumentException("You Have Not Borrow This Book!");
+            }
+        } catch (ClassNotFoundException e) {
+            e.printStackTrace();
+        }
+    }
+
+    private boolean haveRecord(int userId, int bookId) {
         try {
             Connection connect = getConnection();
             String checkSql = "SELECT * FROM Record WHERE user_id = ? AND book_id = ?";
@@ -98,5 +66,45 @@ public class RecordDAOImpl extends BaseDao implements RecordDAO {
             e.printStackTrace();
         }
         return false;
+    }
+
+    @Override
+    public ArrayList<Record> checkOneRecords(int userId) throws IllegalArgumentException {
+        try {
+            Connection connect = getConnection();
+            String checkSql = "SELECT * FROM Record WHERE user_id = ?";
+            PreparedStatement check = connect.prepareStatement(checkSql);
+            check.setObject(1, userId);
+            ResultSet rst = check.executeQuery();
+            if (rst.next()) {
+                ArrayList<Record> records = new ArrayList<>();
+                records.add(buildRecord(rst));
+                while (rst.next()) {
+                    records.add(buildRecord(rst));
+                }
+                closeAll(connect, check, rst);
+                return records;
+            }
+            else {
+                closeAll(connect, check, rst);
+                throw new IllegalArgumentException("No Borrow Record!");
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    private Record buildRecord(ResultSet rst) {
+        try {
+            int userId = rst.getInt("user_id");
+            int bookId = rst.getInt("book_id");
+            Date borrowDay = new Date(rst.getDate("borrow_day").toString());
+            Date returnDay = new Date(rst.getDate("return_day").toString());
+            return new Record(userId, bookId, borrowDay, returnDay);
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return null;
     }
 }
